# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CareerConnect API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  jobseeker: 'jobseeker',
  recruiter: 'recruiter',
  admin: 'admin',
} as const;

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
  role: RegisterInputRole;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  companyId?: number | null;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  name?: string;
  avatar?: string;
  bio?: string;
  location?: string;
  resumeUrl?: string;
  skills?: string;
}

export interface Job {
  id: number;
  title: string;
  companyId: number;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  companyLogo?: string | null;
  location: string;
  type: string;
  level: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  /** @nullable */
  salary?: string | null;
  featured?: boolean;
  active?: boolean;
  /** @nullable */
  postedById?: number | null;
  createdAt?: string;
  /** @nullable */
  applicationCount?: number | null;
}

export interface JobInput {
  title: string;
  companyId: number;
  location: string;
  type: string;
  level: string;
  description: string;
  requirements?: string;
  salary?: string;
  featured?: boolean;
  postedById?: number;
}

export interface JobUpdate {
  title?: string;
  location?: string;
  type?: string;
  level?: string;
  description?: string;
  requirements?: string;
  salary?: string;
  featured?: boolean;
  active?: boolean;
}

export interface JobListResponse {
  jobs: Job[];
  total: number;
  page: number;
  limit: number;
}

export interface Company {
  id: number;
  name: string;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  founded?: number | null;
  /** @nullable */
  openPositions?: number | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  createdAt?: string;
}

export interface CompanyInput {
  name: string;
  logo?: string;
  description?: string;
  website?: string;
  industry?: string;
  size?: string;
  location?: string;
  founded?: number;
}

export interface CompanyUpdate {
  name?: string;
  logo?: string;
  description?: string;
  website?: string;
  industry?: string;
  size?: string;
  location?: string;
  founded?: number;
}

export interface Application {
  id: number;
  jobId: number;
  userId: number;
  status: string;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  companyLogo?: string | null;
  /** @nullable */
  applicantName?: string | null;
  /** @nullable */
  applicantEmail?: string | null;
  /** @nullable */
  applicantAvatar?: string | null;
  appliedAt?: string;
  updatedAt?: string;
}

export interface ApplicationInput {
  jobId: number;
  userId: number;
  coverLetter?: string;
}

export interface ApplicationUpdate {
  status?: string;
}

export interface SavedJob {
  id: number;
  jobId: number;
  userId: number;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  companyLogo?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  level?: string | null;
  /** @nullable */
  salary?: string | null;
  savedAt?: string;
}

export interface SavedJobInput {
  jobId: number;
  userId: number;
}

export interface Review {
  id: number;
  companyId: number;
  userId: number;
  rating: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  pros?: string | null;
  /** @nullable */
  cons?: string | null;
  recommend?: boolean;
  /** @nullable */
  reviewerName?: string | null;
  createdAt?: string;
}

export interface ReviewInput {
  companyId: number;
  userId: number;
  rating: number;
  title?: string;
  pros?: string;
  cons?: string;
  recommend?: boolean;
}

export interface ResumeAnalysisInput {
  resumeText: string;
  jobDescription?: string;
}

export interface ResumeAnalysis {
  score: number;
  strengths: string[];
  improvements: string[];
  keywords: string[];
  summary: string;
}

export interface CoverLetterInput {
  resumeText: string;
  jobDescription: string;
  jobTitle: string;
  companyName: string;
}

export interface CoverLetterResult {
  coverLetter: string;
}

export interface CareerCoachInput {
  question: string;
  context?: string;
}

export interface CareerCoachResult {
  answer: string;
}

export interface InterviewPrepInput {
  jobTitle: string;
  industry: string;
  level?: string;
}

export interface InterviewQuestion {
  question: string;
  tips: string;
  category?: string;
}

export interface InterviewPrepResult {
  questions: InterviewQuestion[];
}

export interface PlatformStats {
  totalUsers: number;
  totalJobs: number;
  totalCompanies: number;
  totalApplications: number;
  newUsersThisMonth?: number;
  newJobsThisMonth?: number;
  activeJobs?: number;
}

export interface DashboardSummary {
  applicationCount: number;
  savedJobCount: number;
  pendingCount?: number;
  interviewCount?: number;
  recentApplications: Application[];
}

export interface RecruiterDashboard {
  postedJobCount: number;
  totalApplicationsReceived: number;
  pendingReviewCount?: number;
  recentApplications: Application[];
}

export type ListJobsParams = {
search?: string;
location?: string;
type?: string;
level?: string;
companyId?: number;
page?: number;
limit?: number;
};

export type ListFeaturedJobsParams = {
limit?: number;
};

export type ListCompaniesParams = {
search?: string;
industry?: string;
page?: number;
limit?: number;
};

export type ListTopCompaniesParams = {
limit?: number;
};

export type ListApplicationsParams = {
userId?: number;
jobId?: number;
status?: string;
};

export type ListSavedJobsParams = {
userId?: number;
};

export type ListReviewsParams = {
companyId?: number;
userId?: number;
};

export type GetDashboardParams = {
userId?: number;
};

export type GetRecruiterDashboardParams = {
userId?: number;
};