# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareerConnect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      parameters:
        - in: query
          name: userId
          schema:
            type: integer
        - in: query
          name: jobId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    patch:
      operationId: updateApplication
      tags: [applications]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
  /ai/analyze-resume:
    post:
      operationId: analyzeResume
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeAnalysisInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeAnalysis"
...
  /ai/career-coach:
    post:
      operationId: careerCoach
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerCoachInput"
      responses:
        "200":
          description: Coaching response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerCoachResult"
  /ai/interview-prep:
    post:
      operationId: interviewPrep
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewPrepInput"
      responses:
        "200":
          description: Interview prep
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewPrepResult"
...
components:
  schemas:
...
    Application:
      type: object
      required: [id, jobId, userId, status]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        userId:
          type: integer
        status:
          type: string
        coverLetter:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        companyLogo:
          type: ["string", "null"]
        applicantName:
          type: ["string", "null"]
        applicantEmail:
          type: ["string", "null"]
        applicantAvatar:
          type: ["string", "null"]
        appliedAt:
          type: string
        updatedAt:
          type: string
    ApplicationInput:
      type: object
      required: [jobId, userId]
      properties:
        jobId:
          type: integer
        userId:
          type: integer
        coverLetter:
          type: string
    ApplicationUpdate:
      type: object
      properties:
        status:
          type: string
...
    ResumeAnalysisInput:
      type: object
      required: [resumeText]
      properties:
        resumeText:
          type: string
        jobDescription:
          type: string
    ResumeAnalysis:
      type: object
      required: [score, strengths, improvements, keywords, summary]
      properties:
        score:
          type: integer
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        summary:
          type: string
...
    CareerCoachInput:
      type: object
      required: [question]
      properties:
        question:
          type: string
        context:
          type: string
    CareerCoachResult:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
    InterviewPrepInput:
      type: object
      required: [jobTitle, industry]
      properties:
        jobTitle:
          type: string
        industry:
          type: string
        level:
          type: string
    InterviewPrepResult:
      type: object
      required: [questions]
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/InterviewQuestion"
    InterviewQuestion:
      type: object
      required: [question, tips]
      properties:
        question:
          type: string
        tips:
          type: string
        category:
          type: string
...