# lib/api-zod/src/generated/types/careerCoachInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CareerConnect API
 * OpenAPI spec version: 0.1.0
 */

export interface CareerCoachInput {
  question: string;
  context?: string;
}